Multifunctional_Calculator.c
This is a multifunctional calculator programmed in C language. 

It provides the following functionalities:

1. Basic Calculator: Perform basic arithmetic operations like addition, subtraction, multiplication, and division.

2. Triangle Math: Calculate the area and perimeter of a triangle.

3. Square Math: Calculate the area and perimeter of a square.

4. Circle Math: Calculate the area and circumference of a circle.

5. Temperature Converter: Convert between Celsius, Fahrenheit, and Kelvin scales.

6. Fibonacci Series: Generate Fibonacci series up to a specified number of terms.

7. Factorial Number: Calculate the factorial of a given number.


This calculator is developed using such a C compiler that supports the <conio.h> header file. 
Therefore, if you want to use it, you need to use a compiler that supports the <conio.h> Header File.